+++ src/main/java/eu/martin/store/product/ProductBuyRequest.java
@@ -3,9 +3,11 @@ package eu.martin.store.product;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ProductBuyRequest(
         @NotNull @Min(1) Short quantity,
-        @NotNull Float buyPrice,
+        @NotNull BigDecimal buyPrice,
         String supplier
 ) {
 }
